% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ACMCalculator-package.R
\docType{package}
\name{ACMCalculator}
\alias{ACMCalculator-package}
\alias{ACMCalculator}
\title{ACMCalculator: A graphical user interface for the analysis of excess deaths in countries.}
\description{
The ACMCalculator package makes it easy to run the WHO WPRO online-calculator
for excess deaths in countries
}
\author{
\strong{Maintainer}: Mark S. Handcock \email{handcock@stat.ucla.edu}

Other contributors:
\itemize{
  \item Bart Blackburn \email{bartblackburn@ucla.edu} [contributor]
}

}
