% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run.R
\name{run}
\alias{run}
\title{Run ACMCalculator}
\usage{
run()
}
\description{
Runs the ACMCalculator shiny application, a GUI for the
WHO WPRO online-calculator for excess deaths in countries
}
\examples{
\dontrun{
run()
}


}
\concept{demography GUI shiny epidemiology}
\keyword{datagen}
\keyword{graphs}
\keyword{models}
